[scheduler:jobs] Refactor job classes to make them extensible
The rationale of this commit is to make the classes needed
to run a job extensible, so other types of jobs can be
created easily. So far, the only type of job available was
'PercevalJob' but other types of jobs will be added in the
near future.

We abandon the idea of having specific job classes and we
back to the suggested RQ workflow which is to create
functions for each job. However, we have defined the new
class 'GrimoireLabJob', that adds extra functionality to
the original RQ 'Job' class. It adds log handlers so any
log message generated by the job will be stored within
the job. Also, a 'progress' property offers the current
progress and status of the job. Therefore, it will be easier
to show the users where the job is at.

We have deleted the 'PercevalJob' class replacing it by the
function 'chronicler_job'. This function contains all the
logic to fetch and eventize data using Perceval and Chronicler.

The JobLogHandler and PercevalResult have been modified too
to make them work with the class and function described
above.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>