[models] Add base model
The class BaseModel will serve as a common database model
for all the other database components under GrimoireLab Core.
It provides the field `created_at`, which stores the creation
date of the entity on the database, and `updated_at`, which
stores when entities are updated. Both fields are managed by
the model and they don't need to be set by the user of the class.

This model has been moved from the scheduler models file.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>