[scheduler:model] Add Task and Job model
These two models represent the way the scheduler will keep track
of recurrent work on the platform. Both classes are Django
abstract classes, which means they can't be instantiated and
they won't have a table on the database.

The idea is that classes to fetch or analyze data will inherit
from the Task class. A new class will be registered using the
function `register_task_model`, that will created automatically
a job model class for that type of task model.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>