[scheduler:db] Add function to look for jobs
The function 'find_job' comes in handy due to the way
Django is able handle models than inherit from an
abstract model. There's no way to use an abstract class
to look for all the entities than inherit from it.
For example, having DummyJob and AnotherDummyJob classes
that inherit from Job, it's not possible to call
Job.objects.get(id=id) to the job that has that identifier.
We need to iterate through all the job models to find
the job.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>