[scheduler:jobs] Fix overwriting logs error
The logs of a job are overwritten due to JobHandler
class. This class initializes the logs everytime
a GrimoireLab job is instantiated. It means that the
platform instantiates a job that it's already running
in a worker, the logs will be emptied and overwriten.

With this change, the JobHandler will only be created
before a job runs. Other instances won't create it.

Also, the JobHandler won't write anymore in 'meta'
attribute. It will call the job's 'add_log' property
instead.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>