[scheduler:worker] Add custom worker
This custom worker allows to change the state of the
Django models (Task and Job) when a job is about to
run.

The worker will update the state of the models when
the job runs. Otherwise, the platform would need
a thread or process that polls RQ to know when a
job has changed its state.

Another option could have been to add a callback that
that triggers when a job starts but that will require
more custom changes in GrimoireLabJob and a custom
worker too.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>