Archive Events in OpenSearch
This commit implements archiving of events in an OpenSearch backend.

It introduces a new subscriber group that reads from the events stream
and stores events in an OpenSearch instance.

Additionally, the worker pool now supports execution on specific task
queues, enabling different workers to process tasks based on their queue.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>