[dist] Build UI when the package is installed/built
The UI is now built when either `poetry install`
or `poetry build` are called. This ensures that the
UI will be distributed and installed without
extra steps.

This is critical when this git repository is installed as
dependency. Previously, only the Python code was installed.

This commit also modifies the .gitignore file to make
sure that only the UI code is distributed within the
source package.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>