Add SortingHat identity extraction functionality
Include a new pool of workers to extract the identities
from events and store them in SortingHat.

The new command to execute these consumers can be run as `grimoirelab
run collect-identities`. This executes a pool of workers that
read events from the Redis Stream and store the identities from
each event in the SortingHat database.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>