Store identity events in SortingHat
Include a new pool of workers to store identities from events in
SortingHat.

The new command to execute these consumers can be run as `grimoirelab
run store-identities`. This executes a pool of workers that
read events from the Redis Stream and store the identities from
each event in the SortingHat database.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>