Add rollover index to archivist settings
Previously, all events were stored in a single index, which
caused degraded performance in OpenSearch when the index
grew too large.

This change introduces support for defining a rollover policy
in the Archivist configuration. The policy allows indices to
roll over automatically once they reach a specified size,
improving search and indexing performance in OpenSearch.

Rollover is now the default behavior, but inserting into a
single index remains supported if explicitly configured.

New environment variables with the default values are:
- GRIMOIRELAB_ARCHIVIST_ROLLOVER_INDICES = True
- GRIMOIRELAB_ARCHIVIST_ROLLOVER_SIZE = 20gb

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>