Add superuser creation to the setup command
Enable superuser creation directly from the `grimoirelab
admin setup` command. This is useful when running the
platform for the first time, as it avoids the need to
execute an additional command. Particularly when
GrimoireLab runs inside a container.

A new `--no-interactive` option has been added to allow
non-interactive execution. When creating a user in
non-interactive mode, if both the username and password
are provided through environment variables, the command
will automatically create the superuser using those values.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>